+++ app/src/main/java/com/chrisvasqm/cuadramo/view/dialogs/PreviewBottomSheetDialogFragment.kt
@@ -1,4 +1,4 @@
-package com.chrisvasqm.cuadramo.dialogs
+package com.chrisvasqm.cuadramo.view.dialogs
 
 import android.os.Bundle
 import android.view.LayoutInflater
