+++ app/src/main/res/layout/activity_about.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/lightGrey"
-    tools:context=".catalog.CatalogActivity">
+    tools:context=".view.catalog.CatalogActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
