+++ app/src/main/res/layout/activity_sign_in.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorPrimary"
-    tools:context=".signin.SignInActivity">
+    tools:context=".view.signin.SignInActivity">
 
     <TextView
         android:id="@+id/textView2"
